+++ RxSwift/Observables/Implementations/Repeat.swift
@@ -1,0 +1,43 @@
+//
+//  Repeat.swift
+//  RxExample
+//
+//  Created by Krunoslav Zaher on 9/13/15.
+//  Copyright © 2015 Krunoslav Zaher. All rights reserved.
+//
+
+import Foundation
+
+class RepeatElement<Element> : Producer<Element> {
+    let element: Element
+    let scheduler: ImmediateSchedulerType
+    
+    init(element: Element, scheduler: ImmediateSchedulerType) {
+        self.element = element
+        self.scheduler = scheduler
+    }
+    
+    override func run<O : ObserverType where O.E == Element>(observer: O, cancel: Disposable, setSink: (Disposable) -> Void) -> Disposable {
+        let sink = RepeatElementSink(parent: self, observer: observer, cancel: cancel)
+        setSink(sink)
+        return sink.run()
+    }
+}
+
+class RepeatElementSink<O: ObserverType> : Sink<O> {
+    typealias Parent = RepeatElement<O.E>
+    
+    let parent: Parent
+    
+    init(parent: Parent, observer: O, cancel: Disposable) {
+        self.parent = parent
+        super.init(observer: observer, cancel: cancel)
+    }
+    
+    func run() -> Disposable {
+        return self.parent.scheduler.scheduleRecursive(self.parent.element) { e, recurse in
+            self.observer?.on(.Next(e))
+            recurse(e)
+        }
+    }
+}