+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 * Removes `SpinLock`s in disposables in favor of more performant `OSAtomicCompareAndSwap32`.
 * Adds `buffer` operator (version with time and count).
 * Adds `range` operator.
+* Adds `repeat` operator.
 
 ## [2.0.0-alpha.2](https://github.com/ReactiveX/RxSwift/releases/tag/2.0.0-alpha.2)